+++ appsrc/mimage.cpp
@@ -122,7 +122,7 @@ MImageImpl::ptr MImageFactory::imageFromBitmap(const MVector<uint8_t>::ptr &bitm
 
 define_reflectable_class_function(MImageFactory, ffDataFromImage, "args:image,format")
 MVector<uint8_t>::ptr MImageFactory::ffDataFromImage(const MImageImpl::ptr &impl, MImageFileFormat format) {
-    implement_injectable_function((MVector<uint8_t>::ptr), impl, (int)format)
+    implement_injectable_function((MVector<uint8_t>::ptr), impl, format)
     return nullptr;
 }
